Use floating point vars in 01

+++ 01-arithmetic.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int a, b;
+    double a, b;
 
     std::cout << "Введите число a: ";
     std::cin >> a;
@@ -14,7 +14,7 @@ int main()
     std::cout << "a - b = " << a - b << std::endl;
     std::cout << "a * b = " << a * b << std::endl;
     if (b != 0) {
-        std::cout << "a / b = " << (double) a / b << std::endl;
+        std::cout << "a / b = " << a / b << std::endl;
     } else {
         std::cout << "a / b - невозможно (b = 0)" << std::endl;
     }
